Adds more fixtures
