+++ src/components/characterCreation/uniques/__tests__/EstablishmentForm.test.tsx
@@ -1,12 +1,6 @@
 import React from 'react';
 import { screen } from '@testing-library/react';
-import {
-  blankCharacter,
-  mockCharacter2,
-  mockEstablishmentCreator,
-  mockGame5,
-  mockKeycloakUserInfo1,
-} from '../../../../tests/mocks';
+import { blankCharacter, mockCharacter2, mockGame5, mockKeycloakUserInfo1 } from '../../../../tests/mocks';
 import { mockKeycloakStub } from '../../../../../__mocks__/@react-keycloak/web';
 import { renderWithRouter, waitOneTick } from '../../../../tests/test-utils';
 import { InMemoryCache } from '@apollo/client';
@@ -14,6 +8,7 @@ import userEvent from '@testing-library/user-event';
 import EstablishmentForm from '../EstablishmentForm';
 import { mockPlayBookCreatorQueryMaestroD } from '../../../../tests/mockQueries';
 import { Game } from '../../../../@types/dataInterfaces';
+import { mockEstablishmentCreator } from '../../../../tests/fixtures/playbookUniqueCreatorsFixtures';
 
 jest.mock('@react-keycloak/web', () => {
   const originalModule = jest.requireActual('@react-keycloak/web');
