+++ src/tests/fixtures/statModifierFixtures.ts
